requirements.txt
Flask>=2.2,<3.0
python-dotenv>=1.0.0
requests>=2.28.0
mysql-connector-python>=8.0.30
pandas>=2.0.0
numpy>=1.24.0
nltk>=3.8.0
gensim>=4.2.0
scikit-learn>=1.2.0
scipy>=1.10.0
keybert>=0.7.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=1.13.0
PyJWT>=2.6.0
Werkzeug>=2.2.0